Implementet the rest of CRUD, remade POST and added Live temp method

+++ DMIPrivateServices/Model/TemperatureData.cs
@@ -36,7 +36,7 @@ namespace DMIPrivateServices.Model
         {
             
         }
-        public TemperatureData(int id, float temperature, DateTime captureTime)
+        public TemperatureData(int id, double temperature, DateTime captureTime)
         {
             _id = id;
             _temperature = temperature;
